Add feature refresh my user information
