+++ server/graphql/resolvers/index.js
@@ -1,0 +1,9 @@
+import Query from "./Query";
+import Mutation from "./Mutation";
+import User from "./User";
+
+export default {
+  Query,
+  Mutation,
+  User
+}
